+++ packages/core/test/request-builder/v4/get-all-request-builder.spec.ts
@@ -43,7 +43,7 @@ describe('GetAllRequestBuilderV4', () => {
 
     it('is built correctly for nested expands', async () => {
       const expected =
-        '/testination/sap/opu/odata/sap/API_TEST_SRV/A_TestEntity?$format=json';
+        '/testination/sap/opu/odata/sap/API_TEST_SRV/A_TestEntity?$format=json&$expand=to_MultiLink($expand=to_MultiLink1($expand=to_MultiLink2))';
       const actual = await requestBuilder
         .expand(
           TestEntity.TO_MULTI_LINK.expand(
@@ -130,7 +130,7 @@ describe('GetAllRequestBuilderV4', () => {
         {
           query: {
             $expand:
-              "to_SingleLink,to_MultiLink($filter=((to_MultiLink/any(a0:(a0/StringProperty ne 'test')))))"
+              "to_SingleLink,to_MultiLink($filter=((to_MultiLink1/any(a0:(a0/StringProperty ne 'test')))))"
           },
           responseBody: { value: [testEntity] }
         },
@@ -140,7 +140,7 @@ describe('GetAllRequestBuilderV4', () => {
         .expand(
           TestEntity.TO_SINGLE_LINK,
           TestEntity.TO_MULTI_LINK.filter(
-            TestEntityMultiLink.TO_MULTI_LINK.filter(
+            TestEntityMultiLink.TO_MULTI_LINK_1.filter(
               any(TestEntityLvl2MultiLink.STRING_PROPERTY.notEquals('test'))
             )
           )
